Merge to current version

+++ python/src/com/jetbrains/python/inspections/quickfix/RedundantParenthesesQuickFix.java
@@ -1,22 +1,9 @@
-/*
- * Copyright 2000-2014 JetBrains s.r.o.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+// Copyright 2000-2018 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
 package com.jetbrains.python.inspections.quickfix;
 
 import com.intellij.codeInspection.LocalQuickFix;
 import com.intellij.codeInspection.ProblemDescriptor;
+import com.intellij.lang.ASTNode;
 import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.project.Project;
 import com.intellij.psi.PsiElement;
@@ -33,11 +20,13 @@ import org.jetbrains.annotations.NotNull;
 public class RedundantParenthesesQuickFix implements LocalQuickFix {
   private static final Logger LOG = Logger.getInstance(RedundantParenthesesQuickFix.class);
 
+  @Override
   @NotNull
   public String getFamilyName() {
     return PyBundle.message("QFIX.redundant.parentheses");
   }
 
+  @Override
   public void applyFix(@NotNull Project project, @NotNull ProblemDescriptor descriptor) {
     PsiElement element = descriptor.getPsiElement();
     if (element instanceof PyParenthesizedExpression) {
@@ -60,7 +49,15 @@ public class RedundantParenthesesQuickFix implements LocalQuickFix {
     else if (element instanceof PyArgumentList) {
       LOG.assertTrue(element.getParent() instanceof PyClass, "Parent type: " + element.getParent().getClass());
       LOG.assertTrue(((PyArgumentList)element).getArguments().length == 0, "Argument list: " + element.getText());
-      element.delete();
+      final ASTNode nameNode = PyElementGenerator.getInstance(project).createFromText(
+        LanguageLevel.forElement(element), PyClass.class, "class A: pass").getNameNode();
+      if (nameNode != null) {
+        final PsiElement emptyArgList = nameNode.getPsi().getNextSibling();
+        element.replace(emptyArgList);
+      }
+      else {
+        element.delete();
+      }
     }
   }
